Redirect out/err to System.out/System.err, rather than with INHERIT.
`ProcessBuilder.Redirect.INHERIT` uses the file descriptors of the
current *process*. That means it does not take into account changes
made to `System.out` and `System.err` by `System.setOut` and
`System.setErr`.

We now use separate threads that explicitly forward the stdout and
stderr of the process to `System.out` and `System.err`.