Fix #12: Force the server and client of Node.js Com to use IPv4.
Node.js 17 switched from resolving host names to IPv4 by default to
resolving to the order given by the OS. This was an intended change
done in https://github.com/nodejs/node/pull/39987, which
nevertheless caused issues in downstream projects, as reported in
https://github.com/nodejs/node/issues/40537.

scalajs-env-nodejs hit that issue, as the JVM server opened on IPv4
by default (apparently this is what the JVM does), but the client
tried to connect via IPv6 with Node.js 17. We fix the issue by
forcing the use of IPv4 in the Node.js client, as well as on the
JVM server for good measure.